requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
numpy>=1.23.0
langchain>=0.1.14
faiss-cpu>=1.7.4
langchain-google-genai>=0.0.7
transformers>=4.37.0
google-generativeai>=0.3.1

